Add getChildText utility function
Introduces a utility function to extract text from a node's children. This function will help resolve fast refresh warnings by moving logic to a separate file.
