+++ src/pages/BlogEntry/BlogEntry.tsx
@@ -20,6 +20,7 @@ import { CustomMarkdownLink } from '../../components/CustomLink/CustomLink';
 import './BlogEntry.css';
 import SectionHeader from '../../components/SectionHeader/SectionHeader';
 import { HeaderRow } from '../../components/Header/Header';
+import { getChildText } from '../../utils/getChildText';
 
 export interface FetchBlogEntryProps {
   id: string;
@@ -34,20 +35,6 @@ export interface BlogEntryProps {}
 export interface BlogEntryType extends React.FC<BlogEntryProps> {
   loadData: (url?: string) => Promise<void>;
 }
-
-// Move function to a separate file later to resolve fast refresh warning
-export const getChildText = (node: any): string => {
-  if (!node || !node.children || node.children.length === 0) return '';
-  const child = node.children[0];
-  if (typeof child === 'string') return child;
-  if (child && typeof child === 'object' && 'value' in child) {
-    return (child as { value: string }).value;
-  }
-  if (child && typeof child === 'object' && 'children' in child) {
-    return getChildText(child);
-  }
-  return '';
-};
 
 export const BlogEntry = () => {
   const navigate = useNavigate();
